feature_flags_SUITE: Adapt the argument to `rabbit_feature_flags:inject_test_feature_flags()`
Up to RabbitMQ 3.10.x, this function took an application attribute. In
`master`, it takes the feature flags map only.

The testsuite now verifies if the remote node supports
`feature_flags_v2` to determine if it should call the function with the
feature flags map directly or convert it to an application attribute
first.

This fixes some failure of the testsuite in mixed-version cluster
testing.