feature_flags_SUITE: Skip `feature_flags_v2` if unsupported by secondary umbrella
The `feature_flags_v2` test group is only relevant if all nodes support
the `feature_flags_v2` feature flags. When doing mixed-version testing,
we must ensure that both umbrellas support that feature flags. If they
are not, we can skip the entire test group.

While here, add a dot at the end of a comment title.