Disconnect the old upgrade mechanism from RabbitMQ boot
We only support upgrades from RabbitMQ 3.8.10+ after #5202. Therefore,
there is no point in keeping this subsystem around. Let's start with
this change. Follow-up commits will remove the now dead code.

I discovered in the process that the upgrade mechanism is also involved
in the regular clustering management and the support for node
renaming... This patch tries to untangle that a bit:
* non-upgrade-related code from `rabbit_upgrade` is copied to
  `rabbit_mnesia` and `rabbit_mnesia_upgrade`.
* we continue to call `rabbit_version:record_desired/0` where it was
  called before.

V2: Calling `rabbit_version:record_desired/0` was very important, not
    just for upgrades: it created a file in the Mnesia directory and
    that file was not filtered out by `rabbit_mnesia:is_virgin_node/0`.
    Therefore, after a node reset and just a call to
    `rabbit_version:record_desired/0`, the node was not considered
    virgin anymore.

    This is especially important for ram nodes (as opposed to disk node)
    because that's the only file present in the Mnesia directory.
    Therefore, removing `rabbit_version:record_desired/0` broken ram
    nodes. The patch now writes another file from
    `init_db_and_upgrade()` just to satisfy the
    `rabbit_mnesia:is_virgin_node/0` logic and make ram nodes work. The
    file is never read.