CQv1: Fix failure to recover messages in rare cases
When a full recovery was done it was possible to lose messages
for v1 queues when the queues only had a journal file and no
segment files.

In practice it should be a rare event because it requires the
queue (or maybe the node) to crash first and then the vhost or
the node to be restarted gracefully.

(cherry picked from commit 744e66e42ae9fd1d8cecf3700773968d925b98bd)
(cherry picked from commit 2695a663564e813b2fbee2c966350abc3ded7aa2)