Set kernel param prevent_overlapping_partitions to true
This kernel parameter got introduced in Erlang 24.3.
It is set to `false` by default in Erlang 24.
It is set to `true` by default in Erlang 25.

This commit requires Erlang >= 24.3.

As described in commit message
https://github.com/rabbitmq/rabbitmq-server/commit/4bf78d822d7496e03061119f4cb07c0b306e4c03
setting this flag to `true` will prevent global:sync/0 from hanging
in the presence of network failures.

Instead of relying on our own workaround of global:sync/0 being stuck
introduced in
https://github.com/rabbitmq/rabbitmq-server/commit/9fcb31f348590a74fd526333cf881cfbe27241e6
let us instead rely on the official Erlang fix that comes by setting
prevent_overloapping_partitions to true.