Enable a post-#4563 test
When a non-mirrored durable classic queue is hosted on a node
that goes down, prior to #4563 not only was the behaviour
that the queue gets deleted from the rabbit_queue table,
but also that its corresponding bindings get deleted.
The purpose of this test was to make sure that bindings
get also properly deleted from the new rabbit_index_route
table.

Given that the behaviour now changed #4563 we can either
delete this test or - as done in this commit - adapt this test.

(cherry picked from commit 23e7fc860b659f9c5eb3f5c99c1c85df632982ba)