mqtt_processor: Get local port from already available adapter_info
This replaces `rabbit_net:socket_ends/2` which also tries to fetch peer
host and port. Fetching peer can return an error if the socket is
already closed (which used to crash the mqtt connection process). This
can happen after a resource alarm, when the processing of CONNECT frame
was delayed and the client closes the connection meanwhile.

```
** Reason for termination ==
** {{badmatch,{error,einval}},
    [{rabbit_mqtt_processor,process_login,4,
                            [{file,"rabbit_mqtt_processor.erl"},{line,585}]},
     {rabbit_mqtt_processor,process_request,3,
                            [{file,"rabbit_mqtt_processor.erl"},{line,143}]},
     {rabbit_mqtt_processor,process_frame,2,
                            [{file,"rabbit_mqtt_processor.erl"},{line,69}]},
     {rabbit_mqtt_reader,process_received_bytes,2,
                         [{file,"src/rabbit_mqtt_reader.erl"},{line,307}]},
```