Set kernel param prevent_overlapping_partitions to true
This commit partially re-applies the changes from #5442.

This kernel parameter got introduced in Erlang 24.3.
It is set to `false` by default in Erlang 24.
It is set to `true` by default in Erlang 25.

This commit requires Erlang >= 24.3.

As described in commit message
https://github.com/rabbitmq/rabbitmq-server/commit/4bf78d822d7496e03061119f4cb07c0b306e4c03
setting this parameter to `true` will prevent global:sync/0 from hanging
in the presence of network failures.

Note that we still rely on our own workaround of global:sync/0 being stuck
introduced in
https://github.com/rabbitmq/rabbitmq-server/commit/9fcb31f348590a74fd526333cf881cfbe27241e6
for rolling upgrades from a cluster with this parameter set to
`false` because according to the Erlang documentation:
"Also note that this fix has to be enabled on all nodes
in the network in order to work properly."