AMQP 1.0 - make it an option to increaese credit on queue confirm
The AMQP 1.0 plugin grants publishers 65000 credit to the senders by default.
This PR makes the credit given configurable.

The AMQP 1.0 plugin grants credit when the message is received, not when it is
placed into the queue. This means that an AMQP 1.0 sender who does not care
about settlement can overload the receiver process leading to high memory
usage. Now it is configurable to grant credit on confirm.

Also fixes a bug in the amqp10_client where the client can run out of credit
and get stuck because of bad accounting.