Check for vhost's existence every 5 seconds
This fixes a few years old bug where a value meant to be interpreted as
seconds was passed to `timer:send_interval/2`, which interprets it as
milliseconds. That led to checking for each vhost's existence a few
times per second, which is excessive. There is no clearly good value,
but a few seconds seems reasonable for the unlikely event of a vhost
crash.

(cherry picked from commit 3f5238fd54540e98ba7f1496b6d414fa26c7b506)