Fix duplicate credit grant in quorum queue
Prior to this commit, when a consumer NACKed a message with requeue=true
(resulting in a Ra #return{} command) and that message got
dropped or dead-lettered (for example because the delivery-limit
was exceeded), that consumer got too many credits granted,
which could lead to exceeding (and therefore violating) the
configured Prefetch value.
The consumer got credit granted twice:
1. for completing the message, and
2. for returning the message.

The same issue occurs in the scenario where a consumer (or its node)
is DOWN. In that case the consumer got credit granted twice:
1. for completing the message (if delivery limit is exceeded), and
2. for returning the message (so that other live consumers can consume)

This bug has existed since 3.8. It got reported in
https://groups.google.com/g/rabbitmq-users/c/iMcX0oXzURQ

From now on, credit for a consumer is increased in only 3 places:

a message is completed, or
a message is returned, or
a message is requeued using the new #requeue{} Ra command (when
delivery-limit is not configured)

Furthermore, this commit also fixes a somewhat related bug:
When a consumer with checked out messages gets cancelled, and
a new consumer subscribes with the same consumer ID and on the same channel
but with a lower Prefetch value, that new consumer's Prefetch value was
not always respected. This commit fixes this issue by ensuring that
the credit in simple_prefetch mode does not exceed the consumer
Prefetch value.

This commit requires a new Ra machine version v3 which is done in-place
in rabbit_fifo due to the small number of changes compared to v2.

(cherry picked from commit 03659864bbd68ce0edfbc64bfd35baf049dd878c)