Fix logout workflow
The issue was primarily that UAA was
not properly configured. We had to whitelist
the uri used for logout otherwise UAA redirects
to its login page

WIP verify that logout.js works when running in
headless mode. For that we need a docker image
and at the moment, make docker-image is not
working because it is still using old otp 24.0.2

(cherry picked from commit e57b88f2b7f8b0dfc783858a183b35f1604179b6)