Change rabbit_fifo_client get_missing_deliveries to use aux_command
As since QQ v2 we don't ever keep any messages in memory and we need
to read them from the log. The only way to do this is by using an
aux command.

Execute get_checked_out query on local members if possible

This reduces the change of crashing a QQ member during a live upgrade
where the follower does not have the appropriate code in handle_aux