Refactor quorum queue periodic metric emission
Primarily to avoid the transient process that is spawned every
Ra "tick" to query the quorum queue process for information that could
be passed in when the process is spawned.

Refactored to make use of the overview map instead of a custom
tuple.

Also use ets:lookup_element where applicable