Optimise shovel pending queue handling for flow-control
The pending queue can grow up to prefetch count (which is 1000 by
default and can be more) while it is forwarded in chunks every time a
bump_credit grants more credit to the shovel process (which is 200 by
default). This change avoids trying to forward all pending entries just
to put back most of them in the pending queue.