Select the otp source file for the OCI explicitly
When the host erlang is detected, it will include the major and minor
version, which may not match an entry in the select statement in
//packaging/docker-image:otp_source for an arbitrary host
erlang. The easiest thing to do just to update the value with
buildozer to the correct one.

(cherry picked from commit e4748eb9ccdb2da12318e3e30d151f1e0892c09f)