Fix ack crashes when using CMQs with v2
The problem comes from the fact that CQv2 does not store the
MsgId in its index when storing in the v2 store as it is not
necessary... or so I thought. It turns out that CMQ slaves
need it to ack, and CMQs need it for sync. This patch ensures
that a correct MsgId is returned in all cases. The only downside
is that when consuming long queues the slaves will need to read
the messages from disk where they didn't before. This only impacts
"embedded" v2 messages.

(cherry picked from commit f854076aa3d5e68dde626c1720a47c33904ad5d9)