AMQP 1.0: Support the modified outcome
Some client libraries (QPid) will automatically send a disposition
with the 'modified' outcome in response to a client local message TTL
expiry.

To support this case and others we treat 'modified' the same as
'accepted' and simply ack the message back to the queue.

This change also contains some API extensions to the amqp10_client
to better support sending the various delivery states (outcomes).