Remove unnecessary call to priority_queue:len/1
Seems like we are not using it anywhere in our code base.
It's unlikely that it's used somewhere else and even if it is,
the API is backwards compatible - we just pass 0, as if the priority_queue
was empty.

(cherry picked from commit 05ec5e488ff06e6548e36d849ae232ce04aec36c)