Refine and fix AMQP 1.0 modified outcome behaviour
The commit in f15e2e42f7a0adba49c325061481b0e29a6b50fd that effectively maps
the modified outcome to accepted is not correct. Modified is a lot more
like the released outcome. This commit refines the behaviour of the modified
in the following ways:

1. A modified outcome with delivery_failed=true, undeliverable_here=false|undefined
will be rejected with requeue=true;
2. Any other modified outcome will be rejected with requeue=false.

It is worth noting that this isn't completely correct but probably the closest
approximation we can achieve at the moment.

The undeliverable_here field refers to the current link, not the message itself but
as we have no means of filtering which messages get assigned to which consumer links
we instead avoid requeuing it.

Also the case where delivery_failed=false|undefined requires the release of the message
_without_ incrementing the delivery_count. Again this is not something that our queues
are able to do so again we have to reject without requeue.

(cherry picked from commit d5cd97784bd3faa7e61d1a242e6c75b4aa90f553)