Adopt new helpers that assert on plugin presence and state
In many tests, we do not care what is the complete set of plugins
running on a node (or present, or in a specific state). We only
care that a few select ones are running, are of the expected version,
in a certain state, and so on.

So list comparison assertions are counterproductive and lead to
test interference that is difficult to track down. In many cases
we can do more fine grained assertions and ignore the rest of
the plugins present on the node.

References #6289, #6020.

(cherry picked from commit 6ebbfaa0a40a11fda3b887af61b788e420a7322a)
(cherry picked from commit 65d1c452a03a5512af0f051b5e00baee808b09f4)