Move Mnesia-specific code to `rabbit_db*` modules
The idea is to isolate in those modules the code to handle the records
in the database. For now, we only support Mnesia, but there is work in
progress to move to another database engine. Doing this reorganization
of the code will also isolate the changes of this upcoming database
engine switch, making them easier to rebase and review.

In the end, the `rabbit_db*` modules should export functions for each
operations we want to perform. Things like "add this record", "list
records matching that", "update a record using an anonymous function
inside a transaction", etc.

Only the following subsystems are impacted in this commit:
* virtual vhosts
* internal users
* runtime parameters

They were modified together because they depend on each other.