Get counters from stream replicas first (#6442)
Previously we queried the nodes in the alphabetical order of the node
names. That means that the writer could be queried first, which could
lead to confusing output where replicas are ahead of the writer (they
have a higher offset). This change makes sure that we query the writer
last.

(cherry picked from commit d3b01942fa08fffb620d318ff9a3cb6c40922003)