rabbit_feature_flags: Revisit the trick we use to convince Dialyzer
The `rabbit_ff_registry` module is generated and compiled at runtime.
The module available in the tree is "empty". The specs are what the
functions will return once generated at runtime.

In the meantime, we have to use some tricks to convince Dialyzer about
that. Indeed the source code in the tree only returns specific values.
The tricks we used so far are not effective anymore.

This patch is here to rework the tricks to "fool" Dialyzer again.