Fix default queue type handling for single vhost imports
Default queue type logic didn't apply when import was done for a
single vhost. Exported definitions for a single vhost don't contain
`vhost` key, and anyway it's better to use the actual vhost name that
will be used for queue declaration.

And just in case, `vhost` is now stripped from all exported items for
a single vhost (it was still present in `parameters` and `policies`) -
on import it was ignored, but it could have been a source of the bug
similar to the one with the default queue type.

(cherry picked from commit 4bfe352168ac4a43ba57f54d8609cb6318450c3d)

# Conflicts:
#	deps/rabbit/src/rabbit_definitions.erl