rabbit_feature_flags: Support virgin node init with feature flags v2
So far, when a new node is started, the feature flags v1 code is used to
enable all stable feature flags (the default behavior). In the list,
there would be `feature_flags_v2`, so this one would be enabled first
and all others would be using the v2 code path.

If we mark `feature_flags_v2` as required, the v2 code path didn't
handle this so far. This patch adds this missing code. Now, a virgin
node's feature flags can be enabled fully using the v2 code path.

As part of that, th `get_forced_feature_flag_names()` function is moved
from `rabbit_feature_flags` to `rabbit_ff_controller`.