rabbit_feature_flags: Make several fixes to `enable_feature_flag_when_ff_file_is_unwritable` testcase
The testcase was moved to another testsuite but called functions from
its initial testsuite. That code code relies on the name of the
testsuite so it broke.

At the same time, make it depend on the testsuite's feature flag, not
stream queues. The stream queues feature flag will become required at
some point and break the testcase.