Preserve heartbeat error reason
https://groups.google.com/g/rabbitmq-users/c/gRURlk9E6M0

In the above thread, the user experienced a heartbeat send timeout that
was logged as `{inet_error, timeout}`. The only way to know this was
related to heartbeats was to reproduce the issue and do a debug trace.
After this change, the error message will be as follows:

```
2022-12-20 08:57:12.614470-08:00 [error] <0.959.0> closing AMQP connection <0.959.0> (127.0.0.1:52052 -> 127.0.0.1:5672):
2022-12-20 08:57:12.614470-08:00 [error] <0.959.0> {inet_error,{heartbeat_send_error,timeout}}
```

I grepped the code for cases where `inet_error` is caught. The one place
where this happens is compatible with the change in this patch.

(cherry picked from commit c7ea2fb92e3f97a73f40907ced16d7ddecd42e61)
(cherry picked from commit f0fe5d69033a50db0a78d95e97adda7b944122b5)